+++ extensions/capso/src/translate-selected-text.ts
@@ -1,4 +1,5 @@
 import { triggerCapsoAction } from "./utils";
+
 export default async function Command() {
-  await triggerCapsoAction("translateSelectedText", 16, "Translate Selected Text");
+  await triggerCapsoAction("capso://grab/translate-selection", "Translate Selected Text");
 }